RFC split bundler code into multiple submodules
